+++ frontend/src/pages/ChartPage.tsx
