+++ frontend/src/pages/CustomerPage.tsx
@@ -3,16 +3,15 @@ import { useGetData } from "../hooks/data";
 import { Box, CircularProgress, MenuItem, radioClasses, Skeleton, TextField } from "@mui/material";
 import { DataGrid, GridColDef, GridCsvExportMenuItem, GridExportMenuItemProps, gridFilteredSortedRowIdsSelector, GridToolbar, GridToolbarColumnsButton, GridToolbarContainer, GridToolbarDensitySelector, GridToolbarExportContainer, GridToolbarFilterButton, gridVisibleColumnFieldsSelector, useGridApiContext } from "@mui/x-data-grid";
 import { red } from "@mui/material/colors";
-import { CustomToolbar } from "./CustomToolbar";
+import { CustomToolbar } from "../components/CustomToolbar";
 import { read } from "node:fs";
 
-interface Reading {
-  customerID: string;
+interface Person {
+  firstName: string;
+  lastName: string;
+  gender: string;
   id: string;
-  kindOfMeter: string;
-  dateOfReading: number[]; 
-  meterCount: number;      
-  comment: string;
+  birthDate: number[] | null;
 }
 
 export enum DataType {
@@ -20,53 +19,41 @@ export enum DataType {
   Readings = 'readings',
 }
 
-const ReadingPage = () => {
-  const { readingsData, loading } = useGetData();
+const CustomerPage = () => {
+  const { customerData, loading } = useGetData();
   const [filter, setFilter] = useState("");
   const [selectedRows, setSelectedRows] = useState<string[]>([]);
   const handleFilterChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setFilter(event.target.value);
   };
 
-  const filteredData = readingsData.filter((reading) => {
-
-    const formattedDate = reading.dateOfReading
-    ? `${reading.dateOfReading[0]}-${String(reading.dateOfReading[1]).padStart(2, '0')}-${String(reading.dateOfReading[2]).padStart(2, '0')}`
-    : "";
-
+  // Filter the data based on the filter input
+  const filteredData = customerData.filter((person) => {
+    const name = `${person.firstName} ${person.lastName}`;
     return (
-      (reading.comment && reading.comment.toLowerCase().includes(filter.toLowerCase())) ||
-      (reading.kindOfMeter && reading.kindOfMeter.toLowerCase().includes(filter.toLowerCase())) ||
-      (reading.id && reading.id.toLowerCase().includes(filter.toLowerCase())) ||
-      (reading.meterCount !== null && reading.meterCount !== undefined && reading.meterCount.toString().toLowerCase().includes(filter.toLowerCase())) ||
-      formattedDate.includes(filter)
+      name.toLowerCase().includes(filter.toLowerCase()) ||
+      person.gender.toLowerCase().includes(filter.toLowerCase())
     );
-  });
+  }
+  );
 
-  const readingsColumns: GridColDef[] = [
-    { field: "customer.id", headerName: "Customer ID", width: 300, valueGetter: (value, row) => {
-      const customerID = row.customer.id;
-      return customerID
-      }, 
-    },
+  // Define the columns for the DataGrid
+  const customersColumns: GridColDef[] = [
     { field: "id", headerName: "ID", width: 300 },
-    { field: "kindOfMeter", headerName: "Kind of Meter", width: 150 },
-    { field: "dateOfReading", headerName: "Date", width: 150,
-        valueGetter: (value, row) => {
-            const readingDate = row.dateOfReading;
-            if (readingDate) {
-              const year = readingDate[0];
-              const month = String(readingDate[1]).padStart(2, '0');
-              const day = String(readingDate[2]).padStart(2, '0');
-              return `${year}-${month}-${day}`;
-            }
-            return "N/A";
-          },
-        },
-    { field: "meterCount", headerName: "Meter Count", width: 150 },
-    { field: "comment", headerName: "Comment", width: 150 },
-    
+    { field: "firstName", headerName: "First Name", width: 150 },
+    { field: "lastName", headerName: "Last Name", width: 150 },
+    { field: "gender", headerName: "Gender", width: 120 },
+    {
+      field: "birthDate",
+      headerName: "Birth Date",
+      width: 150,
+      valueGetter: (value, row) => {
+        const birthDate = row.birthDate;
+        return birthDate ? birthDate.join("-") : "N/A";
+      },
+    },
   ];
+
 
   if (loading) return (
   <Box sx={{ width: "90%", marginBottom: 10 }}>
@@ -74,7 +61,7 @@ const ReadingPage = () => {
     <Skeleton animation="wave" variant="rectangular" height={600} sx={{bgcolor: '#1f1f1f' }}/>
   </Box>);
 
-  if (readingsData.length === 0) return <p>No data available.</p>;
+  if (customerData.length === 0) return <p>No data available.</p>;
 
   return (
     <Box sx={{ height: 600, width: "90%"}}>
@@ -110,7 +97,7 @@ const ReadingPage = () => {
       {/* DataGrid */}
       <DataGrid
         rows={filteredData}
-        columns={readingsColumns}
+        columns={customersColumns}
         slots={{ toolbar: CustomToolbar }}
         checkboxSelection
         disableRowSelectionOnClick
@@ -123,4 +110,4 @@ const ReadingPage = () => {
   );
 };
 
-export default ReadingPage;
+export default CustomerPage;
